+++ swak/funcflow/loggers/stdout.py
@@ -13,7 +13,6 @@ DEFAULT_FMT = '{asctime:<23s} [{levelname:<8s}] {message} ({name})'
 PID_FMT = '{asctime:<23s} [{levelname:<8s}] {message} ({name} | PID-{process})'
 
 
-# ToDo: Truncate log message via flag?
 class PassThroughStdOut(ArgRepr):
     """Pass-through Logger to stdout with at least one formatted StreamHandler.