+++ swak/pt/types.py
@@ -1,3 +1,4 @@
+from typing import Any
 from collections.abc import Callable, Iterator
 from pandas import DataFrame
 from torch import Tensor, dtype, device
@@ -14,7 +15,7 @@ type Tensors = tuple[Tensor, ...]
 type Dtype = dtype
 type Device = device
 type Drop = Dropout | AlphaDropout
-type Batch = tuple[Tensors, Tensor]
+type Batch = tuple[tuple[Any, ...], Tensor]
 type Batches = Iterator[Batch]