+++ swak/text/__init__.py
@@ -11,13 +11,14 @@ from .interpolate import TemplateRenderer, FormFiller
 from .read import TomlReader, YamlReader
 from .parse import YamlParser
 from .write import TomlWriter, YamlWriter
-from .misc import NotFound
+from .misc import NotFound, LiteralNotFound
 
 __all__ = [
     'TextResourceLoader',
     'TemplateRenderer',
     'FormFiller',
     'NotFound',
+    'LiteralNotFound',
     'TomlReader',
     'YamlReader',
     'YamlParser',