+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.6]
+- Added epoch argument to training data call method
+- Changed return signature of TrainData call method
+- Refactored PyTorch model trainer
+- Added cosine learning-rate scheduler
+
 ## [0.2.5] - 2024-11-10
 - Improved handling of missing parent directory in writers/savers
 - Absorbed utility to drop None fields from into TOML writer